+++ app/src/main/java/woowacourse/movie/model/MovieContent.kt
@@ -1,0 +1,10 @@
+package woowacourse.movie.model
+
+data class MovieContent(
+    val imageId: Int,
+    val title: String,
+    val screeningDate: Date,
+    val runningTime: Int,
+    val synopsis: String,
+    val id: Long = 0,
+)
