+++ app/src/main/java/woowacourse/movie/ui/complete/MovieReservationCompleteActivity.kt
@@ -6,9 +6,9 @@ import android.view.MenuItem
 import android.widget.TextView
 import android.widget.Toast
 import woowacourse.movie.R
+import woowacourse.movie.model.MovieContent
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.model.data.MovieContentsImpl
-import woowacourse.movie.model.data.dto.MovieContent
 import woowacourse.movie.ui.DateUi
 import woowacourse.movie.ui.base.BaseActivity
 import woowacourse.movie.ui.complete.constants.CompleteMovieContentKey
