+++ app/src/main/java/woowacourse/movie/ui/complete/MovieReservationCompleteContract.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.ui.complete
 
+import woowacourse.movie.model.MovieContent
 import woowacourse.movie.model.Ticket
-import woowacourse.movie.model.data.dto.MovieContent
 import woowacourse.movie.ui.base.BaseView
 
 interface MovieReservationCompleteContract {
