+++ app/src/test/java/woowacourse/movie/model/MovieContentsImplTest.kt
@@ -5,7 +5,6 @@ import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import woowacourse.movie.model.data.MovieContentsImpl
-import woowacourse.movie.model.data.dto.Date
 
 class MovieContentsImplTest {
     @BeforeEach
